Orbit color changer
Finished the second button of my interact menu which is for the color of the orbits. I used a asset for  it from the Unity Asset Store. I used a checkbox for transparency - if it is checked the color becomes transparent(invisible). Added two buttons - apply for seting the color and Default for reseting the color.